+++ reductstore/src/storage/entry/io/record_writer.rs
@@ -10,16 +10,14 @@ use crate::storage::storage::{CHANNEL_BUFFER_SIZE, MAX_IO_BUFFER_SIZE};
 use async_trait::async_trait;
 use bytes::Bytes;
 use log::error;
-use reduct_base::bad_request;
 use reduct_base::error::ReductError;
+use reduct_base::{bad_request, internal_server_error};
 use std::io::Seek;
 use std::io::SeekFrom;
 use std::io::Write;
 use std::sync::{Arc, RwLock};
 use std::time::Duration;
-use tokio::sync::mpsc::error::SendError;
 use tokio::sync::mpsc::{channel, Receiver};
-use tokio::time::error::Elapsed;
 
 type Chunk = Result<Option<Bytes>, ReductError>;
 type Rx = Receiver<Chunk>;
@@ -29,16 +27,12 @@ pub(crate) trait WriteRecordContent {
     /// Sends a chunk of the record content.
     ///
     /// Stops the writer if the chunk is an error or None.
-    async fn send(&mut self, chunk: Chunk) -> Result<(), SendError<Chunk>>;
+    async fn send(&mut self, chunk: Chunk) -> Result<(), ReductError>;
 
     #[cfg(test)]
-    fn blocking_send(&mut self, chunk: Chunk) -> Result<(), SendError<Chunk>>;
+    fn blocking_send(&mut self, chunk: Chunk) -> Result<(), ReductError>;
 
-    async fn send_timeout(
-        &mut self,
-        chunk: Chunk,
-        timeout: Duration,
-    ) -> Result<Result<(), SendError<Chunk>>, Elapsed>;
+    async fn send_timeout(&mut self, chunk: Chunk, timeout: Duration) -> Result<(), ReductError>;
 }
 
 /// RecordWriter is responsible for writing the content of a record to the storage.
@@ -106,13 +100,6 @@ impl RecordWriter {
         let record_index = block.get_record(time).unwrap();
         let content_size = record_index.end - record_index.begin;
 
-        let (tx, rx) = channel(CHANNEL_BUFFER_SIZE);
-
-        let mut me = RecordWriter {
-            tx,
-            lazy_write: None,
-        };
-
         let ctx = WriteContext {
             bucket_name,
             entry_name,
@@ -125,13 +112,25 @@ impl RecordWriter {
             task_group,
         };
 
-        if content_size >= MAX_IO_BUFFER_SIZE as u64 {
+        let me = if content_size >= MAX_IO_BUFFER_SIZE as u64 {
+            let (tx, rx) = channel(CHANNEL_BUFFER_SIZE);
             shared_child_isolated(&ctx.task_group.clone(), "write record content", move || {
                 Self::receive(rx, ctx);
             });
+            RecordWriter {
+                tx,
+                lazy_write: None,
+            }
         } else {
-            me.lazy_write = Some((rx, ctx));
-        }
+            // for small records we write the content in the same thread
+            // to avoid the overhead of creating a new task
+            // the channel buffer the whole record content, so we need to limit the buffer size for the smallest chunks (8KB)
+            let (tx, rx) = channel(MAX_IO_BUFFER_SIZE / 8_000);
+            RecordWriter {
+                tx,
+                lazy_write: Some((rx, ctx)),
+            }
+        };
 
         Ok(me)
     }
@@ -209,38 +208,31 @@ impl WriteRecordContent for RecordDrainer {
     async fn send(
         &mut self,
         _chunk: Result<Option<Bytes>, ReductError>,
-    ) -> Result<(), SendError<Chunk>> {
+    ) -> Result<(), ReductError> {
         Ok(())
     }
 
     #[cfg(test)]
-    fn blocking_send(&mut self, _chunk: Chunk) -> Result<(), SendError<Chunk>> {
+    fn blocking_send(&mut self, _chunk: Chunk) -> Result<(), ReductError> {
         Ok(())
     }
 
-    async fn send_timeout(
-        &mut self,
-        _chunk: Chunk,
-        _timeout: Duration,
-    ) -> Result<Result<(), SendError<Chunk>>, Elapsed> {
-        Ok(Ok(()))
+    async fn send_timeout(&mut self, _chunk: Chunk, _timeout: Duration) -> Result<(), ReductError> {
+        Ok(())
     }
 }
 
 #[async_trait]
 impl WriteRecordContent for RecordWriter {
-    async fn send(
-        &mut self,
-        chunk: Result<Option<Bytes>, ReductError>,
-    ) -> Result<(), SendError<Chunk>> {
+    async fn send(&mut self, chunk: Result<Option<Bytes>, ReductError>) -> Result<(), ReductError> {
         let stop = chunk.is_err() || chunk.as_ref().unwrap().is_none();
-        self.tx.send(chunk).await?;
+        self.tx.send(chunk).await.map_err(|err| {
+            internal_server_error!("Failed to write the record to internal buffer: {:?}", err)
+        })?;
 
         if stop {
             if let Some((rx, ctx)) = self.lazy_write.take() {
-                // tokio::spawn(async move {  Self::receive(rx, ctx); });
                 tokio::task::spawn_blocking(|| Self::receive(rx, ctx));
-                // Self::receive(rx, ctx);
             }
             self.tx.closed().await;
         }
@@ -249,9 +241,11 @@ impl WriteRecordContent for RecordWriter {
     }
 
     #[cfg(test)]
-    fn blocking_send(&mut self, chunk: Chunk) -> Result<(), SendError<Chunk>> {
+    fn blocking_send(&mut self, chunk: Chunk) -> Result<(), ReductError> {
         let stop = chunk.is_err() || chunk.as_ref().unwrap().is_none();
-        self.tx.blocking_send(chunk)?;
+        self.tx.blocking_send(chunk).map_err(|err| {
+            internal_server_error!("Failed to write the record to internal buffer: {:?}", err)
+        })?;
 
         if stop {
             if let Some((rx, ctx)) = self.lazy_write.take() {
@@ -262,12 +256,12 @@ impl WriteRecordContent for RecordWriter {
         Ok(())
     }
 
-    async fn send_timeout(
-        &mut self,
-        chunk: Chunk,
-        timeout: Duration,
-    ) -> Result<Result<(), SendError<Chunk>>, Elapsed> {
-        tokio::time::timeout(timeout, self.send(chunk)).await
+    async fn send_timeout(&mut self, chunk: Chunk, timeout: Duration) -> Result<(), ReductError> {
+        tokio::time::timeout(timeout, self.send(chunk))
+            .await
+            .map_err(|_| {
+                internal_server_error!("Timeout while writing the record to internal buffer")
+            })?
     }
 }
 
@@ -409,6 +403,29 @@ mod tests {
             );
         }
 
+        #[rstest]
+        #[tokio::test]
+        async fn test_timeout(block_manager: Arc<RwLock<BlockManager>>, block_ref: BlockRef) {
+            let mut writer =
+                RecordWriter::try_new(block_manager.clone(), block_ref, SMALL_RECORD_TIME).unwrap();
+
+            let result = async {
+                // we overload the channel buffer
+                for _ in 0..100 {
+                    writer
+                        .send_timeout(Ok(Some(Bytes::from("x"))), Duration::from_millis(10))
+                        .await?;
+                }
+
+                Ok::<(), ReductError>(())
+            };
+
+            assert_eq!(
+                result.await.err().unwrap(),
+                internal_server_error!("Timeout while writing the record to internal buffer")
+            );
+        }
+
         #[fixture]
         fn path() -> PathBuf {
             tempdir().unwrap().into_path().join("bucket").join("entry")
@@ -455,10 +472,20 @@ mod tests {
 
         #[rstest]
         #[tokio::test]
-        async fn test_does_nothing() {
+        async fn test_send() {
             let mut drainer = RecordDrainer::new();
             drainer.send(Ok(Some(Bytes::from("test")))).await.unwrap();
             drainer.send(Ok(None)).await.unwrap();
+        }
+
+        #[rstest]
+        #[tokio::test]
+        async fn test_send_timeout() {
+            let mut drainer = RecordDrainer::new();
+            drainer
+                .send_timeout(Ok(Some(Bytes::from("test"))), Duration::from_millis(10))
+                .await
+                .unwrap();
         }
     }
 }